reverted changes of mainnet and avalanche

+++ contracts/aggregator/avalanche/flashloan/interfaces.sol
@@ -25,22 +25,18 @@ interface ListInterface {
 interface TokenInterface {
     function approve(address, uint256) external;
 
-    function transfer(address, uint256) external;
+    function transfer(address, uint) external;
 
-    function transferFrom(
-        address,
-        address,
-        uint256
-    ) external;
+    function transferFrom(address, address, uint) external;
 
     function deposit() external payable;
 
-    function withdraw(uint256) external;
+    function withdraw(uint) external;
 
     function balanceOf(address) external view returns (uint256);
 
     function decimals() external view returns (uint256);
-
+    
     function totalSupply() external view returns (uint256);
 }
